+++ tests/unit/test_subscriptions.py
@@ -1,32 +1,46 @@
 """Tests for code relating to the subscriptions."""
 
 ##############################################################################
+# Pytest imports.
+from pytest import mark
+
+##############################################################################
 # Local imports.
-from oldas import Subscription, Subscriptions
+from oldas import Folder, Folders, Subscription, Subscriptions
+from oldas.prefixes import Prefix
+from oldas.subscriptions import Category
+
+##############################################################################
+RAW_TEST_CATEGORY = {"id": Folders.full_id("Test"), "label": "Test"}
+TEST_CATEGORY = Category.from_json(RAW_TEST_CATEGORY)
+
+##############################################################################
+TEST_FOLDER = Folder.from_json({"id": Folders.full_id("Test"), "sortid": ""})
+
 
 ##############################################################################
 TEST_SUBSCRIPTIONS = Subscriptions(
     [
         Subscription.from_json(
             {
-                "id": "1",
+                "id": f"{Prefix.FEED}foo",
                 "title": "z",
                 "sortid": "",
                 "firstitemmsec": 0,
                 "url": "https://example.com/",
                 "htmlUrl": "https://example.com/",
-                "categories": [],
+                "categories": [RAW_TEST_CATEGORY],
             }
         ),
         Subscription.from_json(
             {
-                "id": "1",
+                "id": f"{Prefix.FEED}bar",
                 "title": "a",
                 "sortid": "",
                 "firstitemmsec": 0,
                 "url": "https://example.com/",
                 "htmlUrl": "https://example.com/",
-                "categories": [],
+                "categories": [RAW_TEST_CATEGORY],
             }
         ),
     ]
@@ -41,6 +55,24 @@ def test_sort_subscriptions() -> None:
         "a",
         "z",
     ]
+
+
+##############################################################################
+@mark.parametrize(
+    "data, result",
+    (
+        (TEST_CATEGORY, True),
+        (TEST_CATEGORY.id, True),
+        (TEST_FOLDER, True),
+        (TEST_FOLDER.id, True),
+        ("unknown", False),
+    ),
+)
+def test_subscription_categories_contains(
+    data: str | Category | Folder, result: bool
+) -> None:
+    """We should be able to `in` test categories of a subscription."""
+    assert (data in TEST_SUBSCRIPTIONS[0].categories) is result
 
 
 ### test_subscriptions.py ends here