+++ ChangeLog.md
@@ -1,5 +1,12 @@
 # OldAS ChangeLog
 
+## Unreleased
+
+**Released: WiP**
+
+- Added `in` support to `Subscriptions.Categories` for easier checking.
+  ([#32](https://github.com/davep/oldas/pull/32))
+
 ## v0.7.0
 
 **Released: 2026-01-27**